backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = ""
authors = ["Ashot Svazyan <ashot.svazyan@yandex.ru>"]
readme = "README.md"
repository = "https://github.com/thewitcher1991/vision-net"


[tool.poetry.dependencies]
python = ">=3.12,<3.15"
fastapi = "^0.115.0"
uvicorn = { version = "^0.32.0", extras = ["standard"] }
python-multipart = "^0.0.12"
torch = "^2.10.0"
Pillow = "^10.0.0"
numpy = "^2.4.3"
requests = "^2.32.0"
torchvision = "^0.25.0"
scipy = "^1.17.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^6.0.0"


[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
